pyproject.toml
[project]
name = "vampire-guide"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Team VampireGuide"},
]
dependencies = [
    "osmnx>=1.9.3",
    "folium>=0.17.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.5.0",
    "shapely>=2.0.4",
    "geopandas>=0.14.4",
    "pybdshadow>=0.3.5",
    "haversine>=2.8.1",
    "matplotlib>=3.9.0",
    "mpire>=2.10.2",
    "fastapi>=0.111.0",
    "pydantic>=2.7.4",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[project.optional-dependencies]
dev = [
    "mypy>=1.10.0",
    "pandas-stubs>=2.2.2.240603",
]
[tool.pdm]
distribution = false
